Ok, so, better "rendering", position is now the center of each object, collisions have been rewrite, should work properly, might need some tweak...

+++ Game/VS_Project/TerraSkweek/TerraSkweek/Grid.h
@@ -54,7 +54,7 @@ public:
 	void SetMap(int, int, int);
 	void LoadAllTextures();
 	int LoadGLTextures(string);
-	void PrintImg(int, int, int, int, int);
+	void PrintImg(float, float, float, float, int);
 	void DisplayMap();
 
 	void DrawSpecialCases();
